requirements.txt
requests
beautifulsoup4
langchain
langchain-chroma
langchain-ollama
chromadb
ollama
numpy
python-dotenv
pydantic
ddgs
